Fix/navigation drawer active states (#2430)
* init commit

* fix(v-navigation-drawer): improved model interaction and functionality

updated logic to always prioritize a provided model. if undefined or null is provided, the drawer
will determine if to open based upon the props used, **permanent** or **temporary** or nothing.
changed **enable-resize-watcher** to implicit with a new **disable-resize-watcher** prop. added
better support when switching between different `v-navigation-drawer` types.

fixes #2314

* fix: allow absolute navigation drawer to match content height

* refactor(v-navigation-drawer): reduced function length for codebeat issue

* fix: remove overlay on state change if not temporary

also cleaned up logic and removed some unnecessary code

* Revert "fix: remove overlay on state change if not temporary"

This reverts commit a6fe2476a92e64ba639bd700983bf7557ec3dd87.

* fix: hide overlay when resizing from small to big

if drawer is not a temporary one, it should remove overlay when resizing