feat: added color and opacity props to overlayable mixin (#8861)
There was no way to control the opacity or color of the overlay on VBottomSheet, VDialog and
VNavigationDrawer and the overlayable mixin was apparently only used for the hide-overlay prop. This
adds overlay-color and overlay-opacity props. Also added the translation strings file for the mixin
and adjusted component props on the docs.

fix #8301