docs(contribution): improve info & format (#16082)
## Description
This PR improves the documentation for Contributing. Both in the way
that it improves the information, but it also improves the style /
format of the page at some places of the page.

## Motivation and Context
This is a needed pull request because it will increase the amount of
contributors and make it easier for new contributors to make their first
contribution.

## How Has This Been Tested?
Since this is just documentation, it has been tested locally on my
computer, both with a dev server and after building.

## Markup:
N/A

## Types of changes
<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Improvement/refactoring (non-breaking change that doesn't add any
features but makes things better)

## Checklist:
<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!-- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->
- [x] The PR title is no longer than 64 characters.
- [x] The PR is submitted to the correct branch (`master` for bug fixes
and documentation updates, `dev` for new features and backwards
compatible changes and `next` for non-backwards compatible changes).
- [x] My code follows the code style of this project.
- [x] I've added relevant changes to the documentation (applies to new
features and breaking changes in core library)